More robust solution for sbt script that can't be found or run
The previous solution was a dodgy match against the error that was produced. Now we
check that the sbt script file is executable each time we try to start sbt. This is not going
to cover all bases since some things are executable but not runnable (e.g., symlinks)
but it's better than before.